// proto/generated/core.ts
export interface GetBlockchainStatusRequest {
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBlockchainStatusResponse
 */
export interface GetBlockchainStatusResponse {
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Version version = 1
     */
    version?: GetBlockchainStatusResponse_Version;
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Time time = 2
     */
    time?: GetBlockchainStatusResponse_Time;
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Status status = 3
     */
    status: GetBlockchainStatusResponse_Status;
    /**
     * @generated from protobuf field: double sync_progress = 4
     */
    syncProgress: number;
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Chain chain = 5
     */
    chain?: GetBlockchainStatusResponse_Chain;
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Network network = 7
     */
    network?: GetBlockchainStatusResponse_Network;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Version
 */
export interface GetBlockchainStatusResponse_Version {
    /**
     * @generated from protobuf field: uint32 protocol = 1
     */
    protocol: number;
    /**
     * @generated from protobuf field: uint32 software = 2
     */
    software: number;
    /**
     * @generated from protobuf field: string agent = 3
     */
    agent: string;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBlockchainStatusResponse.Time
 */
export interface GetBlockchainStatusResponse_Time {
    /**
     * @generated from protobuf field: uint32 now = 1
     */
    now: number;
    /**
     * @generated from protobuf field: int32 offset = 2
     */
    offset: number;
    /**
     * @generated from protobuf field: uint32 median = 3
     */
    median: number;
}
// ...
export enum GetBlockchainStatusResponse_Status {
    /**
     * @generated from protobuf enum value: NOT_STARTED = 0;
     */
    NOT_STARTED = 0,
    /**
     * @generated from protobuf enum value: SYNCING = 1;
     */
    SYNCING = 1,
    /**
     * @generated from protobuf enum value: READY = 2;
     */
    READY = 2,
    /**
     * @generated from protobuf enum value: ERROR = 3;
     */
    ERROR = 3
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetMasternodeStatusRequest
 */
export interface GetMasternodeStatusRequest {
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetMasternodeStatusResponse
 */
export interface GetMasternodeStatusResponse {
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.GetMasternodeStatusResponse.Status status = 1
     */
    status: GetMasternodeStatusResponse_Status;
    /**
     * @generated from protobuf field: bytes pro_tx_hash = 2
     */
    proTxHash: Uint8Array;
    /**
     * @generated from protobuf field: uint32 pose_penalty = 3
     */
    posePenalty: number;
    /**
     * @generated from protobuf field: bool is_synced = 4
     */
    isSynced: boolean;
    /**
     * @generated from protobuf field: double sync_progress = 5
     */
    syncProgress: number;
}
// ...
export interface GetBlockRequest {
    /**
     * @generated from protobuf oneof: block
     */
    block: {
        oneofKind: "height";
        /**
         * @generated from protobuf field: uint32 height = 1
         */
        height: number;
    } | {
        oneofKind: "hash";
        /**
         * @generated from protobuf field: string hash = 2
         */
        hash: string;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBlockResponse
 */
export interface GetBlockResponse {
    /**
     * @generated from protobuf field: bytes block = 1
     */
    block: Uint8Array;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBestBlockHeightRequest
 */
export interface GetBestBlockHeightRequest {
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetBestBlockHeightResponse
 */
export interface GetBestBlockHeightResponse {
    /**
     * @generated from protobuf field: uint32 height = 1
     */
    height: number;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.BroadcastTransactionRequest
 */
export interface BroadcastTransactionRequest {
    /**
     * @generated from protobuf field: bytes transaction = 1
     */
    transaction: Uint8Array;
    /**
     * @generated from protobuf field: bool allow_high_fees = 2
     */
    allowHighFees: boolean;
    /**
     * @generated from protobuf field: bool bypass_limits = 3
     */
    bypassLimits: boolean;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.BroadcastTransactionResponse
 */
export interface BroadcastTransactionResponse {
    /**
     * @generated from protobuf field: string transaction_id = 1
     */
    transactionId: string;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetTransactionRequest
 */
export interface GetTransactionRequest {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
}
// ...
export interface BlockHeadersWithChainLocksResponse {
    /**
     * @generated from protobuf oneof: responses
     */
    responses: {
        oneofKind: "blockHeaders";
        /**
         * @generated from protobuf field: org.dash.platform.dapi.v0.BlockHeaders block_headers = 1
         */
        blockHeaders: BlockHeaders;
    } | {
        oneofKind: "chainLock";
        /**
         * @generated from protobuf field: bytes chain_lock = 2
         */
        chainLock: Uint8Array;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.BlockHeaders
 */
export interface BlockHeaders {
    /**
     * @generated from protobuf field: repeated bytes headers = 1
     */
    headers: Uint8Array[];
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetEstimatedTransactionFeeRequest
 */
export interface GetEstimatedTransactionFeeRequest {
    /**
     * @generated from protobuf field: uint32 blocks = 1
     */
    blocks: number;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.GetEstimatedTransactionFeeResponse
 */
export interface GetEstimatedTransactionFeeResponse {
    /**
     * @generated from protobuf field: double fee = 1
     */
    fee: number;
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.TransactionsWithProofsRequest
 */
export interface TransactionsWithProofsRequest {
    /**
     * @generated from protobuf field: org.dash.platform.dapi.v0.BloomFilter bloom_filter = 1
     */
    bloomFilter?: BloomFilter;
    /**
     * @generated from protobuf oneof: from_block
     */
    fromBlock: {
        oneofKind: "fromBlockHash";
        /**
         * @generated from protobuf field: bytes from_block_hash = 2
         */
        fromBlockHash: Uint8Array;
    } | {
        oneofKind: "fromBlockHeight";
        /**
         * @generated from protobuf field: uint32 from_block_height = 3
         */
        fromBlockHeight: number;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: uint32 count = 4
     */
    count: number;
    /**
     * @generated from protobuf field: bool send_transaction_hashes = 5
     */
    sendTransactionHashes: boolean;
}
// ...
export interface TransactionsWithProofsResponse {
    /**
     * @generated from protobuf oneof: responses
     */
    responses: {
        oneofKind: "rawTransactions";
        /**
         * @generated from protobuf field: org.dash.platform.dapi.v0.RawTransactions raw_transactions = 1
         */
        rawTransactions: RawTransactions;
    } | {
        oneofKind: "instantSendLockMessages";
        /**
         * @generated from protobuf field: org.dash.platform.dapi.v0.InstantSendLockMessages instant_send_lock_messages = 2
         */
        instantSendLockMessages: InstantSendLockMessages;
    } | {
        oneofKind: "rawMerkleBlock";
        /**
         * @generated from protobuf field: bytes raw_merkle_block = 3
         */
        rawMerkleBlock: Uint8Array;
    } | {
        oneofKind: undefined;
    };
}
// ...
export interface MasternodeListRequest {
}
/**
 * @generated from protobuf message org.dash.platform.dapi.v0.MasternodeListResponse
 */
export interface MasternodeListResponse {
    /**
     * @generated from protobuf field: bytes masternode_list_diff = 1
     */
    masternodeListDiff: Uint8Array;
}
// ...
export const GetBlockchainStatusRequest = new GetBlockchainStatusRequest$Type();
// ...
export const GetBlockchainStatusResponse = new GetBlockchainStatusResponse$Type();
// ...
export const GetBlockchainStatusResponse_Version = new GetBlockchainStatusResponse_Version$Type();
// ...
export const GetBlockchainStatusResponse_Time = new GetBlockchainStatusResponse_Time$Type();
// ...
export const GetBlockchainStatusResponse_Chain = new GetBlockchainStatusResponse_Chain$Type();
// ...
export const GetMasternodeStatusRequest = new GetMasternodeStatusRequest$Type();
// ...
export const GetMasternodeStatusResponse = new GetMasternodeStatusResponse$Type();
// ...
export const GetBlockRequest = new GetBlockRequest$Type();
// ...
export const GetBlockResponse = new GetBlockResponse$Type();
// ...
export const GetBestBlockHeightRequest = new GetBestBlockHeightRequest$Type();
// ...
export const GetBestBlockHeightResponse = new GetBestBlockHeightResponse$Type();
// ...
export const BroadcastTransactionRequest = new BroadcastTransactionRequest$Type();
// ...
export const BroadcastTransactionResponse = new BroadcastTransactionResponse$Type();
// ...
export const GetTransactionRequest = new GetTransactionRequest$Type();
// ...
class BlockHeadersWithChainLocksResponse$Type extends MessageType<BlockHeadersWithChainLocksResponse> {
    constructor() {
        super("org.dash.platform.dapi.v0.BlockHeadersWithChainLocksResponse", [
            { no: 1, name: "block_headers", kind: "message", oneof: "responses", T: () => BlockHeaders },
            { no: 2, name: "chain_lock", kind: "scalar", oneof: "responses", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
    create(value?: PartialMessage<BlockHeadersWithChainLocksResponse>): BlockHeadersWithChainLocksResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.responses = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<BlockHeadersWithChainLocksResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: BlockHeadersWithChainLocksResponse): BlockHeadersWithChainLocksResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* org.dash.platform.dapi.v0.BlockHeaders block_headers */ 1:
                    message.responses = {
                        oneofKind: "blockHeaders",
                        blockHeaders: BlockHeaders.internalBinaryRead(reader, reader.uint32(), options, (message.responses as any).blockHeaders)
                    };
                    break;
                case /* bytes chain_lock */ 2:
                    message.responses = {
                        oneofKind: "chainLock",
                        chainLock: reader.bytes()
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: BlockHeadersWithChainLocksResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* org.dash.platform.dapi.v0.BlockHeaders block_headers = 1; */
        if (message.responses.oneofKind === "blockHeaders")
            BlockHeaders.internalBinaryWrite(message.responses.blockHeaders, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* bytes chain_lock = 2; */
        if (message.responses.oneofKind === "chainLock")
            writer.tag(2, WireType.LengthDelimited).bytes(message.responses.chainLock);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.dash.platform.dapi.v0.BlockHeadersWithChainLocksResponse
 */
export const BlockHeadersWithChainLocksResponse = new BlockHeadersWithChainLocksResponse$Type();
// ...
export const BlockHeaders = new BlockHeaders$Type();
// ...
export const GetEstimatedTransactionFeeRequest = new GetEstimatedTransactionFeeRequest$Type();
// ...
export const GetEstimatedTransactionFeeResponse = new GetEstimatedTransactionFeeResponse$Type();
// ...
export const TransactionsWithProofsRequest = new TransactionsWithProofsRequest$Type();
// ...
export const TransactionsWithProofsResponse = new TransactionsWithProofsResponse$Type();
// ...
export const MasternodeListRequest = new MasternodeListRequest$Type();
// ...
export const MasternodeListResponse = new MasternodeListResponse$Type();